This merge breaks tabs and one-buffer-one-frame

Conflicts:
	lisp/cus-edit.el
	lisp/mail/reporter.el
	lisp/subr.el
	lisp/window.el

+++ src/nsterm.h
@@ -397,7 +397,7 @@ extern NSMenu *panelMenu;
 #endif
 
 #ifndef NS_HAVE_NSINTEGER
-#if defined(__LP64__) && __LP64__
+#if defined (__LP64__) && __LP64__
 typedef double CGFloat;
 typedef long NSInteger;
 typedef unsigned long NSUInteger;
@@ -460,8 +460,8 @@ struct ns_bitmap_record
 /* this to map between emacs color indices and NSColor objects */
 struct ns_color_table
 {
-  unsigned int size;
-  unsigned int avail;
+  ptrdiff_t size;
+  ptrdiff_t avail;
 #ifdef __OBJC__
   NSColor **colors;
   NSMutableSet *empty_indices;
